Imp(?): added project statistics repos & entities

+++ RR-Backend/src/main/java/com/scor/rr/domain/TargetBuild/Project/Project.java
@@ -24,7 +24,7 @@ public class Project {
     private Integer entity;
 
     @Column(name = "WorkspaceId")
-    private Integer workspaceId;
+    private Long workspaceId;
 
     @Column(name = "ProjectImportRunId")
     private Integer projectImportRunId;
